Restructure cutedsl_megamoe into a shim/ layer behind the package __init__
Collapses the confusing two-layer split (parent nvfp4.py/mxfp8.py wrappers
over megamoe_frontend/api_nvfp4.py/api_mxfp8.py frontends) into a cleaner,
more compact layout that future kernel drops can follow:

  __init__.py    public API for moe_ep; talks only to shim/, re-exports the
                 curated symm-buffer + fused-launch surface (the sole boundary)
  shim/comm.py   dist bootstrap, sym heap, compile state, resolve_gate_up_clamp
                 (was megamoe_frontend/common.py + _util.py)
  shim/nvfp4.py  NVFP4 frontend + symm-buffer/launch wrappers, self-contained
  shim/mxfp8.py  MXFP8 frontend + symm-buffer/launch wrappers, self-contained
  shim/correctness.py  standalone NVFP4 smoke runner

Each dtype is now one self-contained module (frontend + wrapper merged), so
there is no more api_mxfp8.py-vs-mxfp8.py duplication. Path bootstrap moved
into shim/__init__ so the shim is import-safe. The unused ``python -m`` entry
wrappers (__main__.py) were dropped; the smoke runners remain invokable via
``python -m ...cutedsl_megamoe.shim.{nvfp4,mxfp8,correctness}``. SKILL.md
updated for the new layout.
